Major refactor and optimizing.
Major refactor and optimizing for the progress so far on the audio module.  Improved readability and performance. Just have to look over the life time later, since everything is static currently. This will change in the future.  The Audio module is just about complete there are just some core features that are missing.

Going defer Midi, DSP, and HRTF for the future.
Got to really think about if I should implement HRTF, since it quite computationally expensive to compute and is a "nice to have feature".

+++ crates/mani/fabled_audio/src/container/audio_spec.rs
@@ -1,6 +1,6 @@
 use crate::SampleFormat;
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub struct AudioSpecification {
     pub channel_count: u16,
 
@@ -10,8 +10,7 @@ pub struct AudioSpecification {
 
     pub sample_format: SampleFormat,
 
-    // todo maybe convert this to a f32
-    pub duration: u32,
+    pub duration: f32,
 }
 
 impl Default for AudioSpecification {
@@ -21,7 +20,7 @@ impl Default for AudioSpecification {
             sample_rate: 0,
             bit_per_sample: 32,
             sample_format: Default::default(),
-            duration: 0,
+            duration: 0.0,
         }
     }
 }
